added confirmation message before deleting

+++ public/js/mylibrary.js
@@ -6,11 +6,15 @@ $('.deletebtn').on("click", function (event) {
     var selectedItem = $(this);
     var blogId = selectedItem[0].value
     console.log('delete clicked for blog ' + blogId)
+    var confirmDelete = confirm('Are you sure you want to delete this review?');
+    if (!confirmDelete) {
+        return location.reload();
+    }
     fetch((`/api/blogs/${blogId}`),{
         method:"DELETE",
     }).then(res=>{
         if(res.ok){
-            console.log("blog deleted")
+            alert("review deleted!")
             location.href="/mylibrary"
         } else {
             alert("please try again")
